3 important changes: 1- LoggingImpl receives a LoggingSystem and not generic LoggingConsumer 2- HandlerLoggingLevelConfig uses Level class directly instead of ConfigLevel and then bridge to Level. 3- HandlerLoggingLevelConfig is now somehow immutable (not taking cache into account). config updates should create new intances of this class rather calling update method on it. We don't need concurrentHashMaps anymore. 4- remove all defences on HandlerLoggingLevelConfig. Is an internal class and we can control who and how it is called. 5- Tests implementations changed. all config updates now test the loggingSystem rather than HandlerLoggingLevelConfig
