+++ platform-sdk/swirlds-logging-log4j-appender/src/main/java/com/swirlds/logging/log4j/appender/Log4JProvider.java
@@ -20,6 +20,7 @@ import com.swirlds.config.api.Configuration;
 import com.swirlds.logging.api.extensions.event.LogEventConsumer;
 import com.swirlds.logging.api.extensions.event.LogEventFactory;
 import com.swirlds.logging.api.extensions.provider.AbstractLogProvider;
+import com.swirlds.logging.api.internal.LoggingSystem;
 import edu.umd.cs.findbugs.annotations.NonNull;
 
 /**
@@ -49,8 +50,7 @@ public class Log4JProvider extends AbstractLogProvider {
      * @param logEventConsumer the log event consumer
      */
     @Override
-    public void install(
-            @NonNull final LogEventFactory logEventFactory, @NonNull final LogEventConsumer logEventConsumer) {
+    public void install(@NonNull final LogEventFactory logEventFactory, @NonNull final LoggingSystem logEventConsumer) {
         SwirldsLogAppender.setLogEventFactory(logEventFactory);
         SwirldsLogAppender.setLogEventConsumer(logEventConsumer);
     }
