+++ platform-sdk/swirlds-logging/src/main/java/com/swirlds/logging/api/internal/level/HandlerLoggingLevelConfig.java
@@ -16,11 +16,11 @@
 
 package com.swirlds.logging.api.internal.level;
 
+import static com.swirlds.logging.utils.ConfigUtils.configValueOrElse;
+
 import com.swirlds.config.api.Configuration;
 import com.swirlds.logging.api.Level;
 import com.swirlds.logging.api.Marker;
-import com.swirlds.logging.api.extensions.emergency.EmergencyLogger;
-import com.swirlds.logging.api.extensions.emergency.EmergencyLoggerProvider;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import edu.umd.cs.findbugs.annotations.Nullable;
 import java.util.Collections;
@@ -29,7 +29,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.atomic.AtomicReference;
 
 /**
  * This configuration class enables the customization of logging levels for loggers, allowing users to specify levels
@@ -55,88 +54,76 @@ public class HandlerLoggingLevelConfig {
     private static final String PROPERTY_LOGGING_MARKER = "logging.marker";
     private static final String PROPERTY_LOGGING_HANDLER_MARKER = "logging.handler.%s.marker";
     private static final String PROPERTY_LOGGING_HANDLER_INHERIT_LEVELS = "logging.handler.%s.inheritLevels";
-    private static final ConfigLevel DEFAULT_LEVEL = ConfigLevel.INFO;
-
-    /**
-     * The emergency logger.
-     */
-    private static final EmergencyLogger EMERGENCY_LOGGER = EmergencyLoggerProvider.getEmergencyLogger();
+    private static final Level DEFAULT_LEVEL = Level.INFO;
 
     /**
      * The cache for the levels.
      */
-    private final Map<String, ConfigLevel> levelCache;
+    private final Map<String, Level> levelCache = new ConcurrentHashMap<>();
 
     /**
      * The cache for the markers.
      */
-    private final AtomicReference<Map<String, MarkerState>> markerConfigCache;
+    private final Map<String, MarkerState> markerConfigCache;
 
     /**
      * The configuration properties.
      */
-    private final AtomicReference<Map<String, ConfigLevel>> levelConfigProperties;
+    private final Map<String, Level> levelConfigProperties;
 
-    /**
-     * The prefix for the configuration.
-     */
-    private final String handlerName;
+    private HandlerLoggingLevelConfig(
+            Map<String, MarkerState> markerConfigCache, Map<String, Level> levelConfigProperties) {
+        this.markerConfigCache = Collections.unmodifiableMap(markerConfigCache);
+        this.levelConfigProperties = Collections.unmodifiableMap(levelConfigProperties);
+    }
 
     /**
      * Creates a new logging level configuration based on the given configuration and handler name.
      *
      * @param configuration The configuration.
-     * @param handlerName        The name of the handler.
      */
-    public HandlerLoggingLevelConfig(@NonNull final Configuration configuration, @Nullable final String handlerName) {
-        this.handlerName = handlerName;
-        this.levelCache = new ConcurrentHashMap<>();
-        this.markerConfigCache = new AtomicReference<>(new ConcurrentHashMap<>());
-        this.levelConfigProperties = new AtomicReference<>(new ConcurrentHashMap<>());
-        update(configuration);
-    }
-
-    /**
-     * Creates a new root configuration based on the given configuration.
-     * @param configuration The configuration.
-     */
-    public HandlerLoggingLevelConfig(@NonNull final Configuration configuration) {
-        this(configuration, null);
+    public static HandlerLoggingLevelConfig create(final Configuration configuration) {
+        return create(configuration, null);
     }
 
     /**
      * Extracts the configuration for the given handler name based on the given configuration.
      *
-     * @param handlerName The name of the handler.
      * @param configuration The configuration.
+     * @param handlerName   The name of the handler.
      */
     @NonNull
-    private static ExtractedLoggingConfig extractConfig(
-            @Nullable final String handlerName, @NonNull final Configuration configuration) {
+    public static HandlerLoggingLevelConfig create(
+            @NonNull final Configuration configuration, @Nullable final String handlerName) {
         Objects.requireNonNull(configuration, "configuration must not be null");
 
         final ConfigLevel defaultLevel =
-                configuration.getValue(PROPERTY_LOGGING_LEVEL, ConfigLevel.class, ConfigLevel.UNDEFINED);
-        final ConfigLevel defaultHandlerLevel;
+                configValueOrElse(configuration, PROPERTY_LOGGING_LEVEL, ConfigLevel.class, ConfigLevel.UNDEFINED);
+        final Boolean inheritLevels = configValueOrElse(
+                configuration,
+                PROPERTY_LOGGING_HANDLER_INHERIT_LEVELS.formatted(handlerName),
+                Boolean.class,
+                Boolean.TRUE);
+
         final String propertyHandler = PROPERTY_LOGGING_HANDLER_LEVEL.formatted(handlerName);
-        final Boolean inheritLevels = configuration.getValue(
-                PROPERTY_LOGGING_HANDLER_INHERIT_LEVELS.formatted(handlerName), Boolean.class, Boolean.TRUE);
-
-        final Map<String, ConfigLevel> levelConfigProperties = new ConcurrentHashMap<>();
-        final Map<String, MarkerState> markerConfigStore = new ConcurrentHashMap<>();
-
+        final ConfigLevel defaultHandlerLevel;
         if (handlerName != null) {
-            defaultHandlerLevel = configuration.getValue(propertyHandler, ConfigLevel.class, ConfigLevel.UNDEFINED);
+            defaultHandlerLevel =
+                    configValueOrElse(configuration, propertyHandler, ConfigLevel.class, ConfigLevel.UNDEFINED);
         } else {
             defaultHandlerLevel = ConfigLevel.UNDEFINED;
         }
+
+        final Map<String, Level> levelConfigProperties = new HashMap<>();
+        final Map<String, MarkerState> markerConfigStore = new HashMap<>();
+
         if (defaultLevel == ConfigLevel.UNDEFINED && defaultHandlerLevel == ConfigLevel.UNDEFINED) {
             levelConfigProperties.put("", DEFAULT_LEVEL);
         } else if (defaultHandlerLevel != ConfigLevel.UNDEFINED) {
-            levelConfigProperties.put("", defaultHandlerLevel);
+            levelConfigProperties.put("", defaultHandlerLevel.level());
         } else {
             if (Boolean.TRUE.equals(inheritLevels)) {
-                levelConfigProperties.put("", defaultLevel);
+                levelConfigProperties.put("", defaultLevel.level());
             } else {
                 levelConfigProperties.put("", DEFAULT_LEVEL);
             }
@@ -150,19 +137,8 @@ public class HandlerLoggingLevelConfig {
             markerConfigStore.putAll(
                     readMarkers(PROPERTY_LOGGING_HANDLER_MARKER.formatted(handlerName), configuration));
         }
-        return new ExtractedLoggingConfig(levelConfigProperties, markerConfigStore);
-    }
 
-    /**
-     * Updates the handler config based on the given configuration.
-     *
-     * @param configuration The configuration.
-     */
-    public void update(@NonNull final Configuration configuration) {
-        final ExtractedLoggingConfig extractedLoggingConfig = extractConfig(handlerName, configuration);
-        levelCache.clear();
-        this.levelConfigProperties.set(Collections.unmodifiableMap(extractedLoggingConfig.levelConfigProperties()));
-        this.markerConfigCache.set(extractedLoggingConfig.markerConfigStore());
+        return new HandlerLoggingLevelConfig(markerConfigStore, levelConfigProperties);
     }
 
     @NonNull
@@ -184,35 +160,26 @@ public class HandlerLoggingLevelConfig {
     /**
      * Reads the levels from the given configuration.
      *
-     * @param prefix prefix of the configuration property
+     * @param prefix        prefix of the configuration property
      * @param configuration current configuration
-     *
      * @return map of levels and package names
      */
     @NonNull
-    private static Map<String, ConfigLevel> readLevels(
+    private static Map<String, Level> readLevels(
             @NonNull final String prefix, @NonNull final Configuration configuration) {
-        final Map<String, ConfigLevel> result = new HashMap<>();
+        final Map<String, Level> result = new HashMap<>();
         final String fullPrefix = PROPERTY_PACKAGE_LEVEL.formatted(prefix);
 
         configuration.getPropertyNames().filter(n -> n.startsWith(fullPrefix)).forEach(configPropertyName -> {
             final String name = configPropertyName.substring(fullPrefix.length());
             final ConfigLevel level =
                     configuration.getValue(configPropertyName, ConfigLevel.class, ConfigLevel.UNDEFINED);
-            if (level != ConfigLevel.UNDEFINED) {
-                result.put(name, level);
-                // if (containsUpperCase(name)) {
-                // } else {
-                //    result.put(PROPERTY_PACKAGE_LEVEL.formatted(name), level);
-                // }
+            if (level != null && level != ConfigLevel.UNDEFINED) {
+                result.put(name, level.level());
             }
         });
 
         return Collections.unmodifiableMap(result);
-    }
-
-    private static boolean containsUpperCase(@NonNull final String name) {
-        return name.chars().anyMatch(Character::isUpperCase);
     }
 
     /**
@@ -220,25 +187,21 @@ public class HandlerLoggingLevelConfig {
      *
      * @param name  The name of the logger.
      * @param level The level.
-     *
      * @return True if the given level is enabled for the given name.
      */
+    public boolean isEnabled(@NonNull final String name, @NonNull final Level level) {
+        return isEnabled(name, level, null);
+    }
+
     public boolean isEnabled(@NonNull final String name, @NonNull final Level level, @Nullable final Marker marker) {
-        if (level == null) {
-            EMERGENCY_LOGGER.logNPE("level");
-            return true;
-        }
-        if (name == null) {
-            EMERGENCY_LOGGER.logNPE("name");
-            return true;
-        }
+
         if (marker != null) {
+
             final List<String> allMarkerNames = marker.getAllMarkerNames();
             boolean isEnabled = false;
             boolean found = false;
-            final Map<String, MarkerState> stringMarkerStateMap = markerConfigCache.get();
             for (String markerName : allMarkerNames) {
-                final MarkerState markerState = stringMarkerStateMap.get(markerName);
+                final MarkerState markerState = markerConfigCache.get(markerName);
                 if (MarkerState.ENABLED.equals(markerState)) {
                     isEnabled = true;
                     found = true;
@@ -251,15 +214,13 @@ public class HandlerLoggingLevelConfig {
                 return isEnabled;
             }
         }
-
-        final ConfigLevel enabledLevel = levelCache.computeIfAbsent(name.trim(), this::getConfiguredLevel);
+        final Level enabledLevel = levelCache.computeIfAbsent(name, this::getConfiguredLevel);
         return enabledLevel.enabledLoggingOfLevel(level);
     }
 
     @NonNull
-    private ConfigLevel getConfiguredLevel(@NonNull final String name) {
-        final Map<String, ConfigLevel> stringConfigLevelMap = levelConfigProperties.get();
-        ConfigLevel configLevel = stringConfigLevelMap.get(name);
+    private Level getConfiguredLevel(@NonNull final String name) {
+        Level configLevel = levelConfigProperties.get(name);
         if (configLevel != null) {
             return configLevel;
         }
@@ -268,22 +229,12 @@ public class HandlerLoggingLevelConfig {
         for (int i = buffer.length() - 1; i > 0; i--) {
             if ('.' == buffer.charAt(i)) {
                 buffer.setLength(i);
-                configLevel = stringConfigLevelMap.get(buffer.toString());
+                configLevel = levelConfigProperties.get(buffer.toString());
                 if (configLevel != null) {
                     return configLevel;
                 }
             }
         }
-        return stringConfigLevelMap.get("");
+        return levelConfigProperties.get("");
     }
-
-    /**
-     * Helper record to store the extracted logging configuration.
-     *
-     * @param levelConfigProperties The level configuration properties
-     * @param markerConfigStore     The marker configuration store
-     */
-    private record ExtractedLoggingConfig(
-            @NonNull Map<String, ConfigLevel> levelConfigProperties,
-            @NonNull Map<String, MarkerState> markerConfigStore) {}
 }
